add, update, delete, login, http errors

+++ src/main/java/it/synclab/login/service/UserService.java
@@ -5,7 +5,9 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import it.synclab.login.User;
 import it.synclab.login.repository.UserRepository;
@@ -25,19 +27,36 @@ public class UserService {
 	public User getUser(String mail) {
 		Optional<User> obj = userRepository.findById(mail);
 		if (obj.isPresent()) return (obj).get();
-		return null;
+		throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found");
+	}
+	
+	public User loginUser(User user) {
+		Optional<User> obj = userRepository.findById(user.getMail());
+		if (obj.isPresent()) {
+			if (user.getPassword().equals(obj.get().getPassword())) {
+				return (obj).get();
+			} else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Password incorrect");
+		}
+		throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found");
 	}
 	
 	public void addUser(User user) {
+		if(isRegistered(user.getMail())) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Mail already registered"); 
 		userRepository.save(user);
 	}
 
 	public void updateUser(User user) {
+		if(!isRegistered(user.getMail())) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "User not found"); 
 		userRepository.save(user);
 	}
 
 	public void deleteUser(String mail) {
+		if(!isRegistered(mail)) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "User not found"); 
 		userRepository.deleteById(mail);
+	}
+	
+	private boolean isRegistered(String mail) {
+		return userRepository.findById(mail).isPresent();
 	}
 
 }
